requirements.txt
astroid==2.11.2
certifi==2021.10.8
charset-normalizer==2.0.12
dill==0.3.4
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.0
platformdirs==2.5.1
pylint==2.13.4
requests==2.27.1
smmap==5.0.0
tomli==2.0.1
urllib3==1.26.9
wrapt==1.14.0
